fix(scan): rename requests to commands
