+++ packages/scan/src/commands/CreateScan.ts
