+++ packages/scan/src/commands/GetScan.ts
