+++ packages/scan/src/commands/StopScan.ts
