+++ packages/scan/src/commands/index.ts
