+++ packages/scan/src/index.ts
@@ -3,4 +3,4 @@ export * from './Scans';
 export * from './ScanFactory ';
 export * from './HarEntryBuilder';
 export * from './enums';
-export * from './requests';
+export * from './commands';
